+++ src/fat_cache.h
@@ -11,7 +11,7 @@ struct fat_cache {
     off_t partition_offset;
 
     char *cache;
-    int cache_size_blocks;
+    size_t cache_size_blocks;
 
     // Bit field of invalid and dirty bits
     uint8_t *flags;
@@ -24,8 +24,8 @@ struct fat_cache {
 
 int fat_cache_init(struct fat_cache *fc, int fd, off_t partition_offset, size_t cache_size);
 void fat_cache_format(struct fat_cache *fc);
-int fat_cache_read(struct fat_cache *fc, int block, int count, char *buffer);
-ssize_t fat_cache_write(struct fat_cache *fc, int block, int count, const char *buffer);
+int fat_cache_read(struct fat_cache *fc, off_t block, size_t count, char *buffer);
+ssize_t fat_cache_write(struct fat_cache *fc, off_t block, size_t count, const char *buffer);
 ssize_t fat_cache_free(struct fat_cache *fc);
 
 #endif // FAT_CACHE_H
